Addressed bug that affected reconciliation
A scenario where we should've been popping ATM withdrawals from it's
stack was addressed.  The solution would only work in a static
environment, such as the one presented by loading the csv file.  If we
were in a dynamic environment, with new transactions incoming, then we
would have to identify a more elegant solution.  One possible approach
may be utilizing a third Collection, in this case a dequeue.  It would
hold the ATM withdrawals that are too new and have been popped off their
stack.  All subsequent incoming cash transactions could first be checked
against this dequeue. I suspect we would have to take advantage of both
the FIFO and LIFO properties of the dequeue.